Add TTS and ASR module.

+++ app/src/main/java/pku/sei/restaurants/Entry.java
@@ -8,7 +8,7 @@ public class Entry {
 
     public String name = "";
     public int count = 0;
-    public int weight = 0;
+    public double weight = 0;
     public boolean hasEleme = false;
     public boolean hasMeituan = false;
     public boolean hasBaidu = false;
